requirements.txt
stripe==5.1.1
django==2.2.6
python-environ==0.4.54
gunicorn==20.0.4